Refactor out enumerable use result_set instead.
