+++ lib/alma.rb
@@ -2,7 +2,6 @@ require 'alma/version'
 require 'alma/config'
 require 'alma/api_defaults'
 require 'alma/error'
-require 'alma/enumerable'
 require 'alma/alma_record'
 require 'alma/response'
 require 'alma/user'
