+++ lib/alma/fine_set.rb
@@ -1,26 +1,19 @@
+# frozen_string_literal: true
+
 module Alma
-  class FineSet < Alma::Enumerable
-    attr_reader :response
-
-    def_delegators :response, :[], :fetch
-
+  class FineSet < ResultSet
     def key
-      'fee'
+      "fee"
     end
 
     def sum
-      fetch('total_sum', 0)
+      fetch("total_sum", 0)
     end
+
     alias :total_sum :sum
 
     def currency
-      fetch('currency', nil)
+      fetch("currency", nil)
     end
-
-    def total_record_count
-      fetch('total_record_count', 0)
-    end
-
-    alias :total_records :total_record_count
   end
 end
