+++ lib/alma/loan_set.rb
@@ -1,21 +1,15 @@
+# frozen_string_literal: true
+
 module Alma
-  class LoanSet < Alma::Enumerable
-    attr_reader :response
-
-    def_delegators :response, :[], :fetch
-
-    def each
-      @response.fetch(key, []).map{|item| Alma::Loan.new(item)}
-    end
+  class LoanSet < ResultSet
+    alias :total_records :total_record_count
 
     def key
-      'item_loan'
+      "item_loan"
     end
 
-    def total_record_count
-      fetch('total_record_count', 0)
+    def single_record_class
+      Alma::Loan
     end
-
-    alias :total_records :total_record_count
   end
 end
