+++ lib/alma/result_set.rb
@@ -1,42 +1,37 @@
-module Alma
-  class ResultSet < Alma::Enumerable
-    include Alma::Error
+# frozen_string_literal: true
 
-    def_delegators :list, :each, :size
+require "forwardable"
 
-    def total_record_count
-      @response[top_level_key].fetch('total_record_count', 0).to_i
+class Alma::ResultSet
+  extend ::Forwardable
+  include Enumerable
+  include Alma::Error
+
+  attr_reader :response
+
+  def_delegators :response, :[], :fetch
+  def_delegators :each, :each_with_index, :size
+
+  def initialize(response_body_hash)
+    @response = response_body_hash
+  end
+
+  def each
+    @results ||= @response.fetch(key, [])
+      .map { |item| single_record_class.new(item) }
+  end
+
+  def total_record_count
+    fetch("total_record_count", 0).to_i
+  end
+  alias :total_records :total_record_count
+
+  protected
+    def key
+      raise NotImplementedError "Subclasses of ResultSet need to define a response key"
     end
 
-    def list
-      @list ||= list_results
-    end
-
-    def top_level_key
-      raise NotImplementedError 'Subclasses of ResultSet Need to define the top level key'
-    end
-
-    def response_records_key
-      raise NotImplementedError 'Subclasses of ResultSet Need to define the key for response records'
-    end
-
-    private
-    def response_records
-      @response[top_level_key].fetch(response_records_key,[])
-    end
-
-    # Subclasses Can override this to use a Custom Class for single record objects.
     def single_record_class
       Alma::AlmaRecord
     end
-
-    def list_results
-      #If there is only one record in the response, HTTParty returns as a hash, not
-      # an array of hashes, so wrap in array to normalize.
-      response_array = (response_records.is_a? Array) ? response_records : [response_records]
-      response_array.map do |record|
-        single_record_class.new(record)
-      end
-    end
-  end
 end
